extract actions
