+++ src/components/Panel/Actions/index.ts
@@ -1,0 +1,1 @@
+export { Actions } from "./Actions";
